requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
redis==5.0.7
rq==1.16.2
Pillow==10.4.0
openpyxl==3.1.5
python-multipart==0.0.9
requests==2.32.3
